Refactor multi-GPU launcher
